fix : when we go to station page from another station , the poi are displayed
